+++ examples/data-sources/fmc_standard_community_lists/data-source.tf
@@ -1,0 +1,6 @@
+data "fmc_standard_community_lists" "example" {
+  items = {
+    "my_standard_community_lists" = {
+    }
+  }
+}